ExtFUSE: refactor read_args code
Aiming at making the code more clear, rearrange the operation checking
of bpf_extfuse_read_args.
The code should behave the same as before, with the exception of
IN_PARAM_2_SIZE, which is updated to points at the second element of
in.args, instead of the first element as it was before.

Signed-off-by: Alessio Balsini <balsini@android.com>